+++ typescript/registry/src/lib/config.ts
@@ -18,8 +18,6 @@ export const ServerList = [
 ] as const;
 
 export type SDKOptions = {
-  bearerAuth?: string | (() => Promise<string>) | undefined;
-
   httpClient?: HTTPClient;
   /**
    * Allows overriding the default server used by the SDK
@@ -61,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.6.0",
+  sdkVersion: "0.7.0",
   genVersion: "2.743.2",
-  userAgent: "speakeasy-sdk/typescript 0.6.0 2.743.2 1.0.0 @smithery/registry",
+  userAgent: "speakeasy-sdk/typescript 0.7.0 2.743.2 1.0.0 @smithery/registry",
 } as const;